Refactor getDigest
