+++ lib/constructSignatureString.js
@@ -2,17 +2,16 @@
 
 const crypto = require('crypto')
 const parseSignatureHeader = require('./parseSignatureHeader')
-const constructBodyDigestSignature = require('./constructBodyDigestSignature')
 
 const constructSignatureString = (req, options) =>
   calculateSignature({
     algorithm: options.getAlgorithm(),
     sharedSecret: options.sharedSecret,
-    message: getMessage({ req, signature: parseSignatureHeader(req) }),
+    message: getMessage({ req, signature: parseSignatureHeader(req), options }),
     encoding: options.getHMACEncoding()
   })
 
-const getMessage = ({ req, signature }) => {
+const getMessage = ({ req, signature, options }) => {
   const { headers, raw: { method, url } } = req
 
   return signature.headers
@@ -34,7 +33,7 @@ const getMessage = ({ req, signature }) => {
       if (signatureHeader === 'digest') {
         return [
           ...signatureInput,
-          `${signatureHeader}: ${constructBodyDigestSignature(req)}`
+          `${signatureHeader}: ${options.getDigest({ req, options })}`
         ]
       }
       return [
