+++ lib/constructSignatureString.test.js
@@ -2,15 +2,18 @@
 
 const { test } = require('tap')
 const constructSignatureString = require('./constructSignatureString')
+const getDigest = require('./getDigest')
 
 const defaultOptions = {
   sharedSecret: null,
   getAlgorithm: () => {
     throw new Error('No getAlgorithm function provided.')
   },
-  getDigest: () => {
-    throw new Error('No getDigest function provided.')
-  }
+  getHMACEncoding: () => {
+    throw new Error('No getHMACEncoding function provided.')
+  },
+  getDigest,
+  digestEncoding: 'base64'
 }
 const defaultSignature =
   'Nstb3q49bne90WmkbiZ9eyRtKCUmOWvc/kFyw/ftfLtqAXfRZGBk+hSJkXs0GRRm4Q/58c/pdIKj5DND11/fwQ=='
@@ -34,7 +37,7 @@ test('Constructs a signature string from the incoming request', async ({
         ...defaultOptions,
         sharedSecret: 'topSecret',
         getAlgorithm: () => 'sha512',
-        getDigest: () => 'base64'
+        getHMACEncoding: () => 'base64'
       },
       expected: defaultSignature
     },
@@ -55,7 +58,7 @@ test('Constructs a signature string from the incoming request', async ({
         ...defaultOptions,
         sharedSecret: 'topSecret',
         getAlgorithm: () => 'sha512',
-        getDigest: () => 'base64'
+        getHMACEncoding: () => 'base64'
       },
       expected: 'pQul5YFrqv76Zq2bE1kWjJfFGnTu0MwU7X7c8MWDswAI5V7dROqKbBWKUGcoysxujgTqkJo/Eg74x34o54hqRg=='
     }
@@ -77,7 +80,7 @@ test('Throws an error when a signature string cannot be constructed', async ({
       options: {
         ...defaultOptions,
         sharedSecret: 'topSecret',
-        getDigest: () => 'base64'
+        getHMACEncoding: () => 'base64'
       },
       expected: new Error('No getAlgorithm function provided.')
     },
@@ -90,7 +93,7 @@ test('Throws an error when a signature string cannot be constructed', async ({
         sharedSecret: 'topSecret',
         getAlgorithm: () => 'sha512'
       },
-      expected: new Error('No getDigest function provided.')
+      expected: new Error('No getHMACEncoding function provided.')
     }
   ]
