+++ lib/plugin.js
@@ -3,6 +3,7 @@
 const createError = require('http-errors')
 const extractSignature = require('./extractSignature')
 const constructSignatureString = require('./constructSignatureString')
+const getDigest = require('./getDigest')
 
 function plugin (pluginOptions) {
   const defaultOptions = {
@@ -11,11 +12,13 @@ function plugin (pluginOptions) {
     verificationErrorMessage: 'Signature verification failed',
     extractSignature,
     constructSignatureString,
+    getDigest,
+    digestEncoding: 'base64',
     getAlgorithm: () => {
       throw new Error('No getAlgorithm function provided.')
     },
     getHMACEncoding: () => {
-      throw new Error('No getDigest function provided.')
+      throw new Error('No getHMACEncoding function provided.')
     }
   }
